requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
onnxruntime>=1.20.0
numpy>=1.26.0
Pillow>=10.0.0
pytest>=7.4.0
httpx>=0.26.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
requests>=2.31.0
